requirements.txt
discord.py
requests
python-dotenv
flask
apscheduler
pytz
google-generativeai